Added vendor support to com_sales.

+++ com_sales/classes/com_sales.php
@@ -37,6 +37,22 @@ class com_sales extends component {
 	}
 
     /**
+     * Delete a vendor.
+     *
+     * @param int $id The GUID of the vendor.
+     * @return bool True on success, false on failure.
+     */
+	function delete_vendor($id) {
+		if ( $entity = $this->get_vendor($id) ) {
+			$entity->delete();
+            pines_log("Deleted vendor $entity->name.", 'notice');
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+    /**
      * Gets a manufacturer by GUID.
      *
      * @param int $id The manufacturer's GUID.
@@ -44,21 +60,35 @@ class com_sales extends component {
      */
     function get_manufacturer($id) {
         global $config;
-        $manufacturer = $config->entity_manager->get_entity($id);
-        if (is_null($manufacturer) || !$manufacturer->has_tag('com_sales', 'manufacturer'))
-            $manufacturer = null;
-        return $manufacturer;
+        $entity = $config->entity_manager->get_entity($id);
+        if (is_null($entity) || !$entity->has_tag('com_sales', 'manufacturer'))
+            $entity = null;
+        return $entity;
     }
 
     /**
-     * Creates and attaches a module which lists users.
+     * Gets a vendor by GUID.
+     *
+     * @param int $id The vendor's GUID.
+     * @return entity|null The vendor if it exists, null if it doesn't.
+     */
+    function get_vendor($id) {
+        global $config;
+        $entity = $config->entity_manager->get_entity($id);
+        if (is_null($entity) || !$entity->has_tag('com_sales', 'vendor'))
+            $entity = null;
+        return $entity;
+    }
+
+    /**
+     * Creates and attaches a module which lists manufacturers.
      */
 	function list_manufacturers() {
 		global $config;
 
 		$pgrid = new module('system', 'pgrid.default', 'content');
         $pgrid->icons = true;
-        
+
 		$module = new module('com_sales', 'list_manufacturers', 'content');
 		$module->title = "Manufacturers";
         if (isset($_SESSION['user']) && is_array($_SESSION['user']->pgrid_saved_states))
@@ -70,6 +100,29 @@ class com_sales extends component {
             $pgrid->detach();
             $module->detach();
             display_notice("There are no manufacturers.");
+        }
+	}
+
+    /**
+     * Creates and attaches a module which lists vendors.
+     */
+	function list_vendors() {
+		global $config;
+
+		$pgrid = new module('system', 'pgrid.default', 'content');
+        $pgrid->icons = true;
+
+		$module = new module('com_sales', 'list_vendors', 'content');
+		$module->title = "Vendors";
+        if (isset($_SESSION['user']) && is_array($_SESSION['user']->pgrid_saved_states))
+            $module->pgrid_state = $_SESSION['user']->pgrid_saved_states['com_sales/list_vendors'];
+
+		$module->vendors = $config->entity_manager->get_entities_by_tags('com_sales', 'vendor');
+
+		if ( empty($module->vendors) ) {
+            $pgrid->detach();
+            $module->detach();
+            display_notice("There are no vendors.");
         }
 	}
 
@@ -93,6 +146,35 @@ class com_sales extends component {
             $module->entity = $this->get_manufacturer($id);
             if (is_null($module->entity)) {
                 display_error('Requested manufacturer id is not accessible.');
+                $module->detach();
+                return;
+            }
+		}
+        $module->new_option = $new_option;
+        $module->new_action = $new_action;
+        $module->id = $id;
+	}
+
+    /**
+     * Creates and attaches a module containing a form for editing a vendor.
+     *
+     * If $id is null, or not given, a blank form will be provided.
+     *
+     * @param string $heading The heading for the form.
+     * @param string $new_option The option to which the form will submit.
+     * @param string $new_action The action to which the form will submit.
+     * @param int $id The GUID of the vendor to edit.
+     */
+	function print_vendor_form($heading, $new_option, $new_action, $id = NULL) {
+		global $config;
+		$module = new module('com_sales', 'form_vendor', 'content');
+        $module->title = $heading;
+		if ( is_null($id) ) {
+			$module->entity = new entity;
+		} else {
+            $module->entity = $this->get_vendor($id);
+            if (is_null($module->entity)) {
+                display_error('Requested vendor id is not accessible.');
                 $module->detach();
                 return;
             }
